tools/requirements.txt
requests
requests-cache
python-slugify
pydantic == 2.3.0
